Rename smsdelivery/service to smsdelivery/services
