+++ supertokens_python/recipe/passwordless/smsdelivery/services/twilio/implementation.py
@@ -18,7 +18,7 @@ from typing import Any, Dict, Union
 
 from supertokens_python.ingredients.smsdelivery.service.twilio import (
     GetContentResult, ServiceInterface)
-from supertokens_python.recipe.passwordless.smsdelivery.service.twilio.pless_sms import \
+from supertokens_python.recipe.passwordless.smsdelivery.services.twilio.pless_sms import \
     pless_sms_content
 from supertokens_python.recipe.passwordless.types import \
     TypePasswordlessSmsDeliveryInput
